+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [5.0.5](https://github.com/rizen/pixi-glyphs/compare/v5.0.4...v5.0.5) (2025-12-09)
+
 ### [5.0.4](https://github.com/rizen/pixi-glyphs/compare/v5.0.3...v5.0.4) (2025-11-10)
 
 - Fix: Icon position is not offset by outline the way text is.